refactor: added subfolder for eventmanager in views an components
