+++ frontend/src/components/EventManager/EMEventsView/EMTicketSalesCard.vue
